test(PostSubmissionTests) add unit test for auto comfirm
